+++ src/main/java/com/lowdragmc/photon/client/gameobject/emitter/data/shape/IShape.java
@@ -1,5 +1,6 @@
 package com.lowdragmc.photon.client.gameobject.emitter.data.shape;
 
+import com.lowdragmc.lowdraglib2.client.utils.RenderBufferUtils;
 import com.lowdragmc.lowdraglib2.configurator.IConfigurable;
 import com.lowdragmc.lowdraglib2.gui.ColorPattern;
 import com.lowdragmc.lowdraglib2.registry.ILDLRegisterClient;
@@ -48,9 +49,6 @@ public interface IShape extends IConfigurable, IPersistedSerializable, ILDLRegis
     default void drawGuideLines(PoseStack poseStack, MultiBufferSource bufferSource, float partialTicks, IParticleEmitter emitter, Vector3f position, Vector3f rotation, Vector3f scale) {
         var edges = getGuideLines(emitter, position, rotation, scale);
         if (edges.isEmpty()) return;
-        var tessellator = Tesselator.getInstance();
-        var pose = poseStack.last();
-        var mat = pose.pose();
 
         RenderSystem.enableBlend();
         RenderSystem.disableDepthTest();
@@ -58,23 +56,10 @@ public interface IShape extends IConfigurable, IPersistedSerializable, ILDLRegis
 
         RenderSystem.disableCull();
         RenderSystem.setShader(GameRenderer::getRendertypeLinesShader);
-        var buffer = tessellator.begin(VertexFormat.Mode.LINES, DefaultVertexFormat.POSITION_COLOR_NORMAL);
+        var buffer = Tesselator.getInstance().begin(VertexFormat.Mode.LINES, DefaultVertexFormat.POSITION_COLOR_NORMAL);
         RenderSystem.lineWidth(5);
 
-        for (var edge : edges) {
-            var a = edge.getA();
-            var b = edge.getB();
-            float f = b.x - a.x;
-            float f1 = b.y - a.y;
-            float f2 = b.z - a.z;
-            float f3 = Mth.sqrt(f * f + f1 * f1 + f2 * f2);
-            f /= f3;
-            f1 /= f3;
-            f2 /= f3;
-
-            buffer.addVertex(mat, a.x, a.y, a.z).setColor(ColorPattern.YELLOW.color).setNormal(poseStack.last(), f, f1, f2);
-            buffer.addVertex(mat, b.x, b.y, b.z ).setColor(ColorPattern.YELLOW.color).setNormal(poseStack.last(), f, f1, f2);
-        }
+        RenderBufferUtils.drawEdges(poseStack, buffer, edges, ColorPattern.YELLOW.color);
 
         BufferUploader.drawWithShader(buffer.buildOrThrow());
         RenderSystem.enableDepthTest();